Move base and head into common/model_utils
