+++ src/models/CNNDepthMap/CNNDepthMap-height/q3-depthmapmultiartifactlatefusion-plaincnn-height/src/model.py
@@ -43,73 +43,3 @@ def load_base_cgm_model(model_fpath: str, should_freeze: bool=False) -> models.S
             layer.trainable = False
 
     return beheaded_model
-
-
-def create_base_cnn(input_shape, dropout):
-    model = models.Sequential()
-
-    model.add(layers.Conv2D(filters=32, kernel_size=(3, 3), padding="same", activation="relu", input_shape=input_shape))
-    model.add(layers.Conv2D(filters=32, kernel_size=(3, 3), padding="same", activation="relu"))
-    model.add(layers.MaxPooling2D(pool_size=(2, 2)))
-    if dropout:
-        model.add(layers.Dropout(0.05))
-
-    model.add(layers.Conv2D(filters=64, kernel_size=(3, 3), padding="same", activation="relu"))
-    model.add(layers.Conv2D(filters=64, kernel_size=(3, 3), padding="same", activation="relu"))
-    model.add(layers.MaxPooling2D(pool_size=(2, 2)))
-    if dropout:
-        model.add(layers.Dropout(0.075))
-
-    model.add(layers.Conv2D(filters=128, kernel_size=(3, 3), padding="same", activation="relu"))
-    model.add(layers.Conv2D(filters=128, kernel_size=(3, 3), padding="same", activation="relu"))
-    model.add(layers.MaxPooling2D(pool_size=(2, 2)))
-    if dropout:
-        model.add(layers.Dropout(0.1))
-
-    model.add(layers.Conv2D(filters=128, kernel_size=(3, 3), padding="same", activation="relu"))
-    model.add(layers.Conv2D(filters=128, kernel_size=(3, 3), padding="same", activation="relu"))
-    model.add(layers.MaxPooling2D(pool_size=(2, 2)))
-    if dropout:
-        model.add(layers.Dropout(0.125))
-
-    model.add(layers.Conv2D(filters=256, kernel_size=(3, 3), padding="same", activation="relu"))
-    model.add(layers.Conv2D(filters=256, kernel_size=(3, 3), padding="same", activation="relu"))
-    model.add(layers.MaxPooling2D(pool_size=(2, 2)))
-    if dropout:
-        model.add(layers.Dropout(0.15))
-
-    model.add(layers.Conv2D(filters=256, kernel_size=(3, 3), padding="same", activation="relu"))
-    model.add(layers.Conv2D(filters=256, kernel_size=(3, 3), padding="same", activation="relu"))
-    model.add(layers.MaxPooling2D(pool_size=(2, 2)))
-    if dropout:
-        model.add(layers.Dropout(0.175))
-
-    model.add(layers.Flatten())
-
-    model.add(layers.Dense(1024, activation="relu"))
-    if dropout:
-        model.add(layers.Dropout(0.2))
-
-    model.add(layers.Dense(128, activation="relu"))
-    if dropout:
-        model.add(layers.Dropout(0.25))
-
-    return model
-
-
-def create_head(input_shape, dropout):
-    model = models.Sequential(name="head")
-    model.add(layers.Dense(128, activation="relu", input_shape=input_shape))
-    if dropout:
-        model.add(layers.Dropout(0.2))
-
-    model.add(layers.Dense(64, activation="relu"))
-    if dropout:
-        model.add(layers.Dropout(0.2))
-
-    model.add(layers.Dense(16, activation="relu"))
-    if dropout:
-        model.add(layers.Dropout(0.2))
-
-    model.add(layers.Dense(1, activation="linear"))
-    return model
